backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets>=10.0
edge-tts
groq
openai
mistralai
chromadb
webrtcvad-wheels
silero
numpy
pydantic
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain_mistralai
langchain_groq
sentence-transformers
httpx
pyyaml
httpx-sse
tokenizers
